Make bootstrap skill installable
